+++ rust/operator-binary/src/config.rs
@@ -1,3 +1,4 @@
+use crate::authentication::{STACKABLE_SERVER_TLS_DIR, STACKABLE_TLS_STORE_PASSWORD};
 use snafu::{ResultExt, Snafu};
 use stackable_nifi_crd::{
     NifiCluster, NifiConfigFragment, NifiRole, NifiSpec, NifiStorageConfig, HTTPS_PORT,
@@ -507,7 +508,10 @@ pub fn build_nifi_properties(
     // generated with fixed values in the init container
     properties.insert(
         "nifi.security.keystore".to_string(),
-        "/stackable/keystore/keystore.p12".to_string(),
+        format!(
+            "{keystore_path}/keystore.p12",
+            keystore_path = STACKABLE_SERVER_TLS_DIR
+        ),
     );
     properties.insert(
         "nifi.security.keystoreType".to_string(),
@@ -515,11 +519,14 @@ pub fn build_nifi_properties(
     );
     properties.insert(
         "nifi.security.keystorePasswd".to_string(),
-        "secret".to_string(),
+        STACKABLE_TLS_STORE_PASSWORD.to_string(),
     );
     properties.insert(
         "nifi.security.truststore".to_string(),
-        "/stackable/keystore/truststore.p12".to_string(),
+        format!(
+            "{keystore_path}/truststore.p12",
+            keystore_path = STACKABLE_SERVER_TLS_DIR
+        ),
     );
     properties.insert(
         "nifi.security.truststoreType".to_string(),
@@ -527,7 +534,7 @@ pub fn build_nifi_properties(
     );
     properties.insert(
         "nifi.security.truststorePasswd".to_string(),
-        "secret".to_string(),
+        STACKABLE_TLS_STORE_PASSWORD.to_string(),
     );
     properties.insert(
         "nifi.security.user.login.identity.provider".to_string(),